+++ icewm-1.0/src/wmswitch.cc
@@ -63,11 +63,11 @@ void SwitchWindow::resize() {
 				     : 0);
 
     int const iWidth
-	(max(quickSwitchSmallWindow ? (int) manager->width() * 1/3
-				    : (int) manager->width() * 3/5,
-         max(cTitle ? (int) switchFont->textWidth(cTitle) : 0,
+	(max(quickSwitchSmallWindow ? manager->width() * 1/3
+				    : manager->width() * 3/5,
+         max(cTitle ? switchFont->textWidth(cTitle) : 0U,
 	     fIconCount * (YIcon::sizeLarge + 2 * quickSwitchIMargin) +
-	    (quickSwitchHugeIcon ? YIcon::sizeHuge - YIcon::sizeLarge : 0))));
+	    (quickSwitchHugeIcon ? YIcon::sizeHuge - YIcon::sizeLarge : 0U))));
     int const mWidth(manager->width() * 6/7);
     int const iHeight((quickSwitchHugeIcon ? YIcon::sizeHuge
 					   : YIcon::sizeLarge) +
